app/config/config.py
import os
from dotenv import load_dotenv

load_dotenv()

# Ollama
OLLAMA_BASE_URL = os.getenv("OLLAMA_BASE_URL", "http://localhost:11434")
LLM_MODEL = os.getenv("LLM_MODEL", "llama3.1:8b")
EMBEDDING_MODEL = os.getenv("EMBEDDING_MODEL", "nomic-embed-text")

# Chunking
CHUNK_SIZE = int(os.getenv("CHUNK_SIZE", "2000"))
CHUNK_OVERLAP = int(os.getenv("CHUNK_OVERLAP", "200"))

# Retrieval
TOP_K = int(os.getenv("TOP_K", "5"))
SIMILARITY_THRESHOLD = 0.3  # Minimum score to consider a chunk relevant

# Agent
MAX_REWRITE_ATTEMPTS = int(os.getenv("MAX_REWRITE_ATTEMPTS", "3"))

# Paths
UPLOAD_DIR = os.path.join(os.path.dirname(__file__), "..", "..", "uploads")
FAISS_INDEX_DIR = os.path.join(os.path.dirname(__file__), "..", "..", "faiss_index")
