requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langgraph
chromadb
unstructured
huggingface_hub
sentence-transformers
transformers
pdfplumber
python-docx
unstructured[local]
langchain 
chromadb 
sentence-transformers
python-dotenv

langchain==0.3.10
langchain-community==0.3.10
langchain-ollama==0.3.0
langgraph==0.5.3
faiss-cpu==1.9.0.post1
PyMuPDF==1.25.3
fastapi==0.115.11
uvicorn==0.34.0
streamlit==1.39.0
python-dotenv==1.0.1
pydantic==2.10.4
requests==2.32.3
numpy==1.26.4
